requirements.txt
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.15
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
graphtools==1.5.3
imageio==2.36.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.0.3
phate==1.0.11
pillow==11.0.0
PyGSP==0.5.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
s_gd2==1.8.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scprep==1.2.3
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
tasklogger==1.2.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
wrapt==1.17.0
